add feature flags and remove old methods
