+++ validator/client/propose.go
@@ -11,6 +11,7 @@ import (
 	validatorpb "github.com/prysmaticlabs/prysm/proto/validator/accounts/v2"
 	"github.com/prysmaticlabs/prysm/shared/bls"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
+	"github.com/prysmaticlabs/prysm/shared/featureconfig"
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/timeutils"
 	"github.com/prysmaticlabs/prysm/validator/slashing-protection/remote"
@@ -96,8 +97,15 @@ func (v *validator) ProposeBlock(ctx context.Context, slot uint64, pubKey [48]by
 	if remoteProtector, ok := v.remoteSlashingProtector.(*remote.Service); ok && remoteProtector != nil {
 		remoteSlashable, err := v.remoteSlashingProtector.IsSlashableBlock(ctx, blk, pubKey, signingRoot)
 		if err != nil {
-			// If slasher is unavailable, trust local protection and proceed with submitting the attestation.
-			if !errors.Is(err, remote.ErrSlasherUnavailable) {
+			if featureconfig.Get().DisableStrictRemoteSlashingProtection {
+				if !errors.Is(err, remote.ErrSlasherUnavailable) {
+					// If slasher is unavailable, trust local protection and proceed with submitting the attestation.
+					log.WithFields(
+						blockLogFields(pubKey, blk),
+					).WithError(err).Warn("Could not check block safety with remote slashing protection, not submitting")
+					return
+				}
+			} else {
 				log.WithFields(
 					blockLogFields(pubKey, blk),
 				).WithError(err).Warn("Could not check block safety with remote slashing protection, not submitting")
